test(ma): calendar - add integration and divide components
